requirements.txt
asgiref==3.6.0
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
coverage==7.1.0
decorator==5.1.1
distlib==0.3.6
Django==4.1.6
djangorestframework==3.14.0
drf-jsonmask==0.3.0
drf-yasg==1.21.4
executing==1.2.0
factory-boy==3.2.1
Faker==16.6.1
filelock==3.9.0
identify==2.5.15
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
ipdb==0.13.9
ipython==8.8.0
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
jsonmask==0.1.1
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
nodeenv==1.7.0
packaging==23.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.6.2
pluggy==1.0.0
pre-commit==2.20.0
prompt-toolkit==3.0.36
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
pytest==7.2.0
pytest-cov==4.0.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
ruamel.yaml==0.17.21
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
toml==0.10.2
traitlets==5.8.1
uritemplate==4.1.1
urllib3==1.26.14
virtualenv==20.17.1
wcwidth==0.2.6
